requirements.txt
Flask
openai
nltk
python-dotenv
python-dotenv
email-validator
scikit-learn
joblib
pandas
